Moves getNeighbours into TECrop and also moves the test for it.
